+++ src/components/AdminDashboard/SearchBarAdmin/SearchBarAdmin.module.scss
@@ -1,0 +1,27 @@
+.ctnInput {
+    display: flex;
+    align-items: center;
+    gap: 10px;
+  }
+  
+  .inputSearch {
+    background-color: white;
+    height: 35px;
+    width: 100vw;
+    max-width: 400px;
+    border-radius: 8px;
+    padding-left: 10px;
+  }
+  
+  .btnIconSearch {
+    height: 40px;
+    width: 45px;
+    border-radius: 8px;
+    cursor: pointer;
+    transition: all 0.3s ease-in-out;
+    background-color: rgb(65,64,64);
+  }
+  
+  .btnIconSearch:hover {
+    transform: scale(1.05);
+  }
